refactor(applications): split ApplicationFormActions helper into its own module
react-doctor only-export-components requires component files to export
only components. Move the discriminated-mode helper (and its ActionsMode
type) to ApplicationFormActions.helpers.ts.
